playground: add notebook for exploring pstable LSH
Some interesting findings about norm of uniform random vecotrs bloomed
out of this, which is nice. Otherwise, I've only just begun to learn
about the p-stable property, and its use in LSH.

Also added a script for estimating the std on a normal distribution that
approximates the norm of uniform vectors. I call this distribution M,
and it's similar to the Irwin-Hall distribution, I suppose. I struggled
to find info on it.